added CRUD funcionality

+++ server.py
@@ -27,7 +27,15 @@ def showCatalog():
 # categories CRUD
 @app.route('/catalog/category/new', methods=['GET', 'POST'])
 def newCategory():
-    return render_template('new_category.html')
+    if request.method == 'GET':
+        return render_template('new_category.html')
+    if request.method == 'POST':
+        name = request.form['name']
+        description = request.form['description']
+        new_category = Category(name=name, description=description)
+        session.add(new_category)
+        session.commit()
+        return redirect(url_for('showCatalog'))
 
 
 @app.route('/catalog/<string:category_name>')
@@ -43,7 +51,16 @@ def showCategory(category_name):
 def editCategory(category_name):
     selected_category = session.query(Category).filter_by(name=category_name).first()
     if selected_category:
-        return render_template('edit_category.html', category=selected_category)
+        if request.method == 'GET':
+            return render_template('edit_category.html', category=selected_category)
+        if request.method == 'POST':
+            name = request.form['name']
+            description = request.form['description']
+            selected_category.name = name
+            selected_category.description = description
+            session.add(selected_category)
+            session.commit()
+            return redirect(url_for('showCategory', category_name=name))
     return render_template('404.html')
 
 
@@ -51,15 +68,33 @@ def editCategory(category_name):
 def deleteCategory(category_name):
     selected_category = session.query(Category).filter_by(name=category_name).first()
     if selected_category:
-        return render_template('delete_category.html', category=selected_category)
+        if request.method == 'GET':
+            return render_template('delete_category.html', category=selected_category)
+        if request.method == 'POST':
+            session.delete(selected_category)
+            session.commit()
+            return redirect(url_for('showCatalog'))
     return render_template('404.html')
 
 
 # items CRUD
 @app.route('/catalog/item/new', methods=['GET', 'POST'])
 def newItem():
-    categories = session.query(Category).order_by(Category.name).all()
-    return render_template('new_item.html', categories=categories)
+    if request.method == 'GET':
+        categories = session.query(Category).order_by(Category.name).all()
+        return render_template('new_item.html', categories=categories)
+    if request.method == 'POST':
+        name = request.form['name']
+        description = request.form['description']
+        category_id = request.form['category']
+        quantity = request.form['quantity']
+        new_item = Item(name=name,
+                        description=description,
+                        quantity=quantity,
+                        category_id=category_id)
+        session.add(new_item)
+        session.commit()
+        return redirect(url_for('showItem', category_name=new_item.category.name, item_name=name))
 
 
 @app.route('/catalog/<string:category_name>/<string:item_name>')
@@ -76,7 +111,20 @@ def editItem(category_name, item_name):
     categories = session.query(Category).order_by(Category.name).all()
     selected_item = session.query(Item).filter_by(name=item_name).first()
     if selected_item:
-        return render_template('edit_item.html', categories=categories, item=selected_item)
+        if request.method == 'GET':
+            return render_template('edit_item.html', categories=categories, item=selected_item)
+        if request.method == 'POST':
+            name = request.form['name']
+            description = request.form['description']
+            category_id = request.form['category']
+            quantity = request.form['quantity']
+            selected_item.name = name
+            selected_item.description = description
+            selected_item.category_id = category_id
+            selected_item.quantity = quantity
+            session.add(selected_item)
+            session.commit()
+            return redirect(url_for('showItem', category_name=selected_item.category.name, item_name=name))
     return render_template('404.html')
 
 
@@ -85,7 +133,12 @@ def deleteItem(category_name, item_name):
     selected_category = session.query(Category).filter_by(name=category_name).first()
     selected_item = session.query(Item).filter_by(name=item_name).first()
     if selected_category and selected_item:
-        return render_template('delete_item.html', category=selected_category, item=selected_item)
+        if request.method == 'GET':
+            return render_template('delete_item.html', category=selected_category, item=selected_item)
+        if request.method == 'POST':
+            session.delete(selected_item)
+            session.commit()
+            return redirect(url_for('showCatalog'))
     return render_template('404.html')
